+++ src/components/manage/Blocks/HeroImageLeft/View.test.jsx
@@ -3,7 +3,9 @@ import renderer from 'react-test-renderer';
 import View from './View';
 
 test('renders a view hero component', () => {
-  const component = renderer.create(<View data={{ url: 'heroimage.jpg' }} />);
+  const component = renderer.create(
+    <View data={{ url: 'http://localhost:8080/Plone/heroimage.jpg' }} />,
+  );
   const json = component.toJSON();
   expect(json).toMatchSnapshot();
 });