+++ src/components/manage/Blocks/Listing/DefaultTemplate.jsx
@@ -4,6 +4,7 @@ import { ConditionalLink } from '@plone/volto/components';
 import { flattenToAppURL } from '@plone/volto/helpers';
 import { settings } from '~/config';
 
+import Image from '@plone/volto/components/theme/Image/Image';
 import DefaultImageSVG from '@plone/volto/components/manage/Blocks/Listing/default-image.svg';
 import { isInternalURL } from '@plone/volto/helpers/Url/Url';
 
@@ -34,11 +35,8 @@ const DefaultTemplate = ({ items, linkMore, isEditMode }) => {
                 <img src={DefaultImageSVG} alt="" />
               )}
               {item[settings.listingPreviewImageField] && (
-                <img
-                  src={flattenToAppURL(
-                    item[settings.listingPreviewImageField].scales.preview
-                      .download,
-                  )}
+                <Image
+                  image={item[settings.listingPreviewImageField]}
                   alt={item.title}
                 />
               )}