+++ src/components/theme/View/ImageView.test.jsx
@@ -21,11 +21,44 @@ test('renders an image view component', () => {
           title: 'Hello World!',
           description: 'Hi',
           image: {
+            download: 'http://localhost:8080/Plone/test-images/@@images/image',
             size: 123123,
-            download: 'file:///preview.jpg',
+            width: 1920,
             scales: {
+              icon: {
+                download:
+                  'http://localhost:8080/Plone/test-images/@@images/image/icon',
+                width: 32,
+              },
+              large: {
+                download:
+                  'http://localhost:8080/Plone/test-images/@@images/image/large',
+                width: 768,
+              },
+              listing: {
+                download:
+                  'http://localhost:8080/Plone/test-images/@@images/image/listing',
+                width: 16,
+              },
+              mini: {
+                download:
+                  'http://localhost:8080/Plone/test-images/@@images/image/mini',
+                width: 200,
+              },
               preview: {
-                download: 'file:///preview.jpg',
+                download:
+                  'http://localhost:8080/Plone/test-images/@@images/image/preview',
+                width: 400,
+              },
+              thumb: {
+                download:
+                  'http://localhost:8080/Plone/test-images/@@images/image/thumb',
+                width: 128,
+              },
+              tile: {
+                download:
+                  'http://localhost:8080/Plone/test-images/@@images/image/tile',
+                width: 64,
               },
             },
           },