+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 ### Feature
 
 - Lazy load image in blocks Image and HeroImage @mamico
+- Add ImageFromUrl with srcset and lazy loading using Plone miniatures @nzambello
 
 ### Bugfix